Level 1 Begin Blockout
Made a Level 1 Scene with a different enemy pool of boids and angry boids.

Created 2 new prefabs specifically for this stage and 2 new scripts that make it so that the boids work with the enemy manager AEC system.

Also messed around with some settings stuff, specifically FOV and retention over sessions, aswell as camera fix